### 2022-04-01 #4294 @editorialbot
[PRE REVIEW]: MeMC: A package for monte-carlo simulations of spherical shells
**Submitting author:** @vipinagrawal25 (<a href="http://orcid.org/0000-0002-1291-5035">Vipin Agrawal</a>)
**Repository:** https://github.com/vipinagrawal25/MeMC
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
