### 2022-04-01 #4294 @kyleniemeyer
Hi @vipinagrawal25, one requirement for JOSS submissions is that they are released under an [OSI-approved open source license](https://opensource.org/licenses). Can you please select and add one to your repository?
