### 2022-04-01 #4294 @vipinagrawal25
Dear @kyleniemeyer,

Thanks for your comments. I mainly see two issues right now:

1) I have now added a GPL 3.0 license to the repository.
2) We have a section called "Statement of need and purpose of software" in the paper.

I hope it answers the comments.

Regards
Vipin
