### 2022-04-02 #4294 @kyleniemeyer
@openjournals/dev I'm not sure what is causing this error in building a paper; the error message is
```
/opt/hostedtoolcache/Ruby/3.1.1/x64/lib/ruby/3.1.0/psych.rb:670:in `initialize': No such file or directory @ rb_sysopen - tmp/11b2b7af7bcbdf67ea3877aa28d9207d/doc/paper.yml (Errno::ENOENT)
[95](https://github.com/openjournals/joss-papers/runs/5801744333?check_suite_focus=true#step:3:95)
	from /opt/hostedtoolcache/Ruby/3.1.1/x64/lib/ruby/3.1.0/psych.rb:670:in `open'
[96](https://github.com/openjournals/joss-papers/runs/5801744333?check_suite_focus=true#step:3:96)
	from /opt/hostedtoolcache/Ruby/3.1.1/x64/lib/ruby/3.1.0/psych.rb:670:in `load_file'
```

Could it be this extra block in the paper header?
```yaml
output:
  pdf_document:
    extra_dependencies: "subfig"
    bibliography: paper.bib
```
