+++ doc/Makefile
@@ -1,6 +1,0 @@
-default:
-	@pandoc paper.md -f markdown -t latex -s -o paper.pdf
-tex2md:
-	@pandoc -s paper.tex -o paper.txt
-md2tex:
-	@pandoc paper.md -f markdown -t latex -s -o paper.tex
