### 2022-04-04 #4294 @fboehm
@victorapm, @mgiardino, @kris-rowe, @junghans, @tschwinge - would any of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
