### 2022-04-07 #4294 @mgiardino
Hi @fboehm <https://github.com/fboehm>, I accept reviewing this submission.
Cheers,

Marco

> Il giorno 4 apr 2022, alle ore 21:23, Frederick Boehm ***@***.***> ha scritto:
>
>
> @victorapm <https://github.com/victorapm>, @mgiardino <https://github.com/mgiardino>, @kris-rowe <https://github.com/kris-rowe>, @junghans <https://github.com/junghans>, @tschwinge <https://github.com/tschwinge> - would any of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html <https://joss.readthedocs.io/en/latest/review_criteria.html>
