### 2022-04-21 #4305 @mgiardino
Hi @fboehm, I'm working on it and indeed I have a question(It's my first JOSS review). Besides the checklist that we are following, do we have to produce some sort of report to gather our comments? I have some advice/comments right now that I wish to communicate to the author, but I'm not sure if I'm supposed to create an issue about it or just write it here as a comment.
