### 2022-04-21 #4305 @fboehm
@mgiardino and @victorapm - Thanks for the questions!

You are not required to produce a report with comments. I recognize that such a report may be standard for other journals, but at JOSS we primarily depend on teh checklists. That being said, all comments for improving the package are welcome. You are also welcome to ask questions of the authors. To offer a comment or to ask a question of the authors, please do so in this comment thread. It's helpful, I think, to tag the submitting author if you wish to ask them a question.

If you like, you may also open issues in the submitted software repository (which is distinct from the joss-reviews repo). If you choose this approach, please be sure to link to any opened issues in this thread so that we have a complete record of activity in this thread. Thanks again!
