### 2022-04-26 #4305 @mgiardino
Hi @vipinagrawal25, I have to report an issue about the Prerequisites of your README.
I've tried the proposed instruction for Ubuntu systems in my Ubuntu 20.04.4 LTS:
_apt install g++ hdf5_
resulting in “cannot found packet hdf5”.
Maybe this requirement shall be better specified (which language binding is required? which version? some repo to add?).
For example, some possible packets are: libhdf5-dev, libhdf5-cpp-103.
Anyway, I finally achieved my compilation task by following your proposed solution on using the h5c++ compiler.
