### 2022-04-26 #4305 @mgiardino
One more thing: to successfully execute the python modules, avoiding the error "ModuleNotFoundError: No module named 'h5py'", the _h5py_ module shall be added to the pip installation instruction.
