### 2022-04-28 vipinagrawal25/MeMC#1 @mgiardino
Corrections in README.md
**In section "MeMC"**

- Linked image is not valid (404 error): https://github.com/vipinagrawal25/MeMC/blob/main/doc/figs/describe_theta.png

**In section "Example"**

- These image links are broken:
https://github.com/vipinagrawal25/MeMC/raw/main/doc/figs/hist_start.png
https://github.com/vipinagrawal25/MeMC/raw/main/doc/figs/hist_memc.png

**In section "Data Structure"**

- sentence “the fourth, fifth, sixth and the seventh column (...)”
shall be changed to:
“the fourth, fifth, sixth, seventh and the eighth column (...)”

- in the sentence “we write the positions in snap_xx.h5 series, where xx represents Monte Carlo iterations”
shall be changed to something similar to:
“we write the positions in snap_xx.h5 series, where xx represents Monte Carlo iterations divided by 100”

**In section "Visualization"**

- sentence “If the execution is successful, the file "check_viz.vtk" will be written in the root directory.”
is wrong, as the file names written in the root dir will be viz_000.vtk or viz_010.vtk.

- the Visit link is wrong, maybe it shall be replaced by this one: https://visit-dav.github.io/visit-website/index.html
