### 2022-04-29 #4305 @mgiardino
Hi @fboehm, I've completed the installation steps and I'm now using my running instance to perform several test runs.
With visit I am visualizing the resulting vtks and also trying to analyze them. My comprehension of the involved science domain is very limited, anyway, so my considerations are mostly focused on sw engineering features. About the "Missing DOIs" from editioralbot, are we supposed to use also this information in some way?
Cheers,

Marco
