### 2022-04-29 #4305 @victorapm
@fboehm, it's going well. The comments above regarding installation helped me. Thanks, @mgiardino. Now, I'm looking at reproducing the results.

I have a question: is it necessary that all authors of the paper have at least one commit in the code repo?

Thanks!
