### 2022-05-01 #4305 @fboehm
@mgiardino - Thank you for your question. The missing DOIs should be corrected by the authors, but you need not do anything with them.

@victorapm - Thank you for your question. There is no JOSS policy requiring that authors have at least one commit.

Thank you, @mgiardino and @victorapm for your work on these reviews. @vipinagrawal25 - please ensure that your cited works have accurate dois.
