+++ README.md
@@ -1,8 +1,7 @@
-
 # MeMC
 The MeMC is an open-source software package for monte-carlo simulation of elastic shells. The package can be useful to study the mechanics of biological nano-vesicles e.g. Exosomes. 
 
-Micro and nano vesicles play a crucial role in biology and medicine. The physical properties of these vesicles play an important role in their biological functions. Hence it is important to measure their elastic constants. One of the ways, to measure elastic constants of cells, is to poke them with AFM tip to compute force-distance curve. Then we model cell as a linear elastic material and by fitting this model to the experimental force-distance curve, we estimate the parameters of elastic model i.e. cell . However nano-vesicles differ from cells in two ways:
+Micro and nano vesicles play a crucial role in biology and medicine. The physical properties of these vesicles play an important role in their biological functions. Hence it is important to measure their elastic constants. One of the ways, to measure elastic constants of cells, is to poke them with AFM tip to compute force-distance curve. Then we model cell as a linear elastic material and by fitting this model to the experimental force-distance curve, we estimate the parameters of elastic model i.e. cell. However nano-vesicles differ from cells in two ways:
 
  1) The nano-vesicles are much smaller hence thermal fluctuations may effectively renormalize the elastic coefficients. ([Košmrlj & Nelson, 2017](https://journals.aps.org/prx/abstract/10.1103/PhysRevX.7.011002), [Paulose et al., 2012](https://www.pnas.org/doi/abs/10.1073/pnas.1212268109)).
  2) Cell membranes are strongly coupled to an underlying cytoskelton. Hence they may be modeled by a solid body ([HW et al., 2002](https://www.pnas.org/doi/abs/10.1073/pnas.202617299)) but nano-vesicles must be modeled as liquid filled elastic membranes.
@@ -33,7 +32,7 @@ In such a scenario, the user should set these paths manually in the Makefile to 
 In order to install the required python libraries we suggest using the standard
 python package manager pip, 
 ```bash
-pip install scipy numpy-quaternion
+pip install scipy numpy-quaternion h5py
 ```
 
 ## Installation
@@ -46,8 +45,8 @@ git clone https://github.com/vipinagrawal25/MeMC
 cd MeMC
 make
 ```
-If successful, one should find an extra "bin" directory in the folder. The bin
-directory will contain the binaries "exe_start" and "exe_memc".
+If successful, one should find an extra `bin` directory in the folder. The bin
+directory will contain the binaries `exe_start` and `exe_memc`.
 
 ---
 **NOTE**
@@ -75,19 +74,19 @@ Atomic Force Microscopy. This package tends to simulate the experiment in the be
 possible way. Therefore, we expect the user to have a prior understanding of
 the Metropolis algorithm for the Monte Carlo simulation and a basic concept of Elasticity
 pertaining to membranes. Both the concept is described in detail in the document
-paper/paper.pdf.  That being said, we shall now dive deeper and explain how different
+`paper/paper.pdf`.  That being said, we shall now dive deeper and explain how different
 part of the code functions in the following sections:
 
 ## Constructing the membrane
 
 To begin the simulation we generate a equilibrated randomized position on a surface
-of a sphere.  For details we refer the reader to section xx.xx of paper/paper.pdf. The
-main code for this purpose is given in `main/start.cpp` and the relevant  executable is "bin/exe_start". 
+of a sphere.  For details we refer the reader to `subsection grid` in `section numerical implementation` of `paper/paper.pdf`. The
[…]
