### 2022-05-02 c84b42a
minor edits in readme
