### 2022-05-02 vipinagrawal25/MeMC#1 @vipinagrawal25
@mgiardino ,

Thanks a lot for your nice and elaborate comments. We agree with all of your suggestions and we have changed `README.md` file accordingly.
