### 2022-05-02 #4305 @vipinagrawal25
> Thank you @vipinagrawal25. One more thing: to successfully execute the python modules, avoiding the error "ModuleNotFoundError: No module named 'h5py'", the _h5py_ module shall be added to the pip installation instruction.

Hi @mgiardino, We have now added the instructions for h5py installation in the readme as well.
