### 2022-05-02 #4305 @vipinagrawal25
> Hi @vipinagrawal25 , I've created the issue [vipinagrawal25/MeMC#1](https://github.com/vipinagrawal25/MeMC/issues/1) to suggest some minor corrections to your README.md file. Regards,
>
> Marco

Hi @mgiardino,

Thanks a lot for the comments. We have now incorporated all of them and closed the issue. Do let us know if you have any further suggestions.
