### 2022-05-03 #4305 @mgiardino
Dear @vipinagrawal25, after constructing the membrane, triangulating the surface, running the bin/exe_memc executable, the program gives the following console output message (multiple times): “X0 small roots may not be correct”. I'm not sure if this message shall be seriously taken into consideration, as the results are still being produced. In any case, the output stream is crowded with these messages, maybe this behavior can be confusing for the user.
