### 2022-05-08 640bc80
utility file to check status
