+++ utils/gen_memc_conf.py
@@ -154,7 +154,6 @@ sort_tri = sort_simplices(triangles)
 cmlist, node_nbr = neighbours(Np, sort_tri)
 node_nbr = sort_nbrs(pts_cart, Np, cmlist, node_nbr)
 isDir = os.path.isdir("./conf/") 
-pts_cart[:,:] = 0e0
 if(isDir):
     write_hdf5(pts_cart, cmlist, node_nbr,
              triangles, "./conf/dmemc_pos.h5",
