### 2022-05-08 #4305 @thecalculon
Dear @mgiardino,

The section of code where we evaluate the force exerted by the AFM (Atomic Force
Microscopy) tip to Exosome outputs the message "X0 small roots may not be correct".
Consider the representative configuration of the Exosome the AFM tip, shown in the figure attached.
The AFM tip is modeled as a paraboloid and the force it exerts is very short ranged.
Therefore, we assume that the force exerted on a represented point
say (x_0, y_0, z_0) on exosome is only from a point (x,y,z) of the tip which is
nearest to (x_0, y_0, z_0). Thus, to evaluate the force we first solve for (x,y,z)
in the code.

When the value of x_0 (or y_0/z_0) is zero till machine precision,
we find the solution to be incorrect. At the same time,
if the points representing the surface of Exosome is determined using
a Monte-Carlo simulation then, we have observed that this condition (x_0=0)
is never satisfied. Therefore, we had never seen the message in our screen
as you are observing.

Fortunately, after several attempts on different laptops and servers we were able to
reproduce the error. At our end the reason was that the memc config file
(conf/dmemc_pos.h5) was not written/read properly. We were able to fix the
error by reinstalling the hdf5 library.

To avoid the confusion to the user, we now have changed the error message to the
following;

`
x0 = 0, y0 = 0,  z0 = 0; Points representing exosome very small...
See section Checking execution status in https://github.com/vipinagrawal25/MeMC/blob/main/README.md
`

We have also added a section `Checking Execution Status` in the README file which
should enable the user to debug the code properly.

@mgiardino we hope that we were able to help you resolve the error. We welcome all further
suggestions.

![error](https://user-images.githubusercontent.com/41811674/167305827-0718e3c2-6ef0-4883-84b0-a2d8383c2df3.png)
