### 2022-05-12 #4305 @mgiardino
Hi @fboehm, I think my review is pretty complete. The paper is well written and the software behavior is "as expected": great job!
The authors were really open to my suggestions and I think that the small corrections performed could also simplify the use of this module.
I write here my last comments.

**Statement of need**
The target audience is somewhat implicit.

**References**
The (n.d. …) style references are syntactically correct?

**Functionality/Performance**
No performance claims.

**Documentation/Functionality documentation**
No API or very minimal documentation is provided inline with the source code.
The sw usage is well described, but the code is missing any kind of description, considering file, method, or instruction level. No documentation support is provided for the developer trying to understand or contribute to the software logic.

**Documentation/Community guidelines**
I have not found any clear guidelines about contributions for third parties.
