### 2022-05-15 8adc1f6
minor corrections to Readme
