### 2022-05-15 d661ddd
minor corrections to Readme
