### 2022-05-15 #4305 @thecalculon
> Dear @thecalculon, thank you for your very interesting explanation! I have reinstalled the hfd5 library and downloaded/recompiled the MeMC code. Now the error messages have disappeared and everything seems OK. The new "Checking Execution Status" section is also helpful. Maybe it should also be declared the dependence of the scriptutils/check_status.py from the matplotlib module.

Hi @mgiardino we have updated the Readme.md adding the matplotlib dependency . Since the software is required only when the user is in the debugging phase, we have added the dependency in the section "Checking Execution status"
