+++ paper/paper.bib
@@ -1,4 +1,3 @@
-
 @book{phillips2012physical,
   title={Physical biology of the cell},
   author={Phillips, Rob and Kondev, Jane and Theriot, Julie and Garcia, Hernan G and Orme, Nigel},
@@ -14,7 +13,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={3},
   pages={2628--2636},
   year={2017},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acsnano.6b07302}
 }
 @article{paulose2012fluctuating,
   title={Fluctuating shells under pressure},
@@ -24,7 +24,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={48},
   pages={19551--19556},
   year={2012},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.1212268109}
 }
 
 @article{kovsmrlj2017statistical,
@@ -207,4 +208,3 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
   key={Monte-Carlo for elastic shells},
   howpublished="\url{https://github.com/vipinagrawal25/MeMC}"
 }
-
