+++ paper/paper.bib
@@ -36,7 +36,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={1},
   pages={011002},
   year={2017},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevx.7.011002}
 }
 
 @article{hw2002stomatocyte,
@@ -47,7 +48,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={26},
   pages={16766--16769},
   year={2002},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.202617299}
 }
 @misc{comsol,
   key={COMSOL},
@@ -61,7 +63,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={1},
   pages={221},
   year={1999},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevlett.82.221}
 }
 @article{bowick2001universality,
   title={Universality classes of self-avoiding fixed-connectivity membranes},
@@ -71,7 +74,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={2},
   pages={149--160},
   year={2001},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s101890170071}
 }
 @article{auth2005fluctuation,
   title={Fluctuation spectrum of membranes with anchored linear and star polymers},
@@ -81,7 +85,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={3},
   pages={031904},
   year={2005},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreve.72.031904}
 }
 @inproceedings{gompper2004triangulated,
   title={Triangulated-surface models of fluctuating membranes},
@@ -90,7 +95,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   editor={Nelson, David and Piran, Tsvi and Weinberg, Steven},
   pages={359--426},
   year={2004},
-  publisher={World Scientific}
+  publisher={World Scientific},
+  doi={10.1103/physrevx.7.011002}
 }
 @book{LLelast,
   title={Theory of Elasticity},
@@ -105,7 +111,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   title={Robust and Efficient Delaunay triangulations of points on or close to a sphere},
   author={Caroli, Manuel and de Castro, Pedro Machado Manh{\~a}es and Loriot, Sebastien and Rouiller, Olivier and Teillaud, Monique and Wormser, Camille},
   year={2009},
-  school={INRIA}
+  school={INRIA},
+  doi={10.1007/978-3-642-13193-6_39}
 }
 @article{vliegenthart2006mechanical,
   title={Mechanical deformation of spherical viruses with icosahedral symmetry},
@@ -115,7 +122,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={3},
   pages={834--841},
   year={2006},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1529/biophysj.106.081422}
 }
 @article{buenemann2008elastic,
   title={Elastic properties and mechanical stability of chiral and filled viral capsids},
@@ -125,7 +133,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={5},
   pages={051924},
   year={2008},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreve.78.051924}
 }
 @misc{meshzoo,
   key={A simple and fast mesh generator},
@@ -147,7 +156,8 @@ volume={36},
[…]
