### 2022-05-30 edaac7d
misc references done
