+++ paper/paper.bib
@@ -52,8 +52,10 @@ title={Competition between bending and internal pressure governs the mechanics o
   doi={10.1073/pnas.202617299}
 }
 @misc{comsol,
-  key={COMSOL},
-  howpublished="\url{https://www.comsol.com}"
+  author={COMSOL Inc.},
+  title={COMSOL},
+  year={2020},
+  url={http://www.comsol.com/products/multiphysics/},
 }
 @article{goetz1999mobility,
   title={Mobility and elasticity of self-assembled membranes},
@@ -137,8 +139,11 @@ title={Competition between bending and internal pressure governs the mechanics o
   doi={10.1103/physreve.78.051924}
 }
 @misc{meshzoo,
+  author={Nico Schlömer},
+  title={Meshzoo},
   key={A simple and fast mesh generator},
-  howpublished="\url{https://github.com/meshpro/meshzoo}"
+  year={2020},
+  howpublished="\url{https://github.com/meshpro/meshzoo/releases}"
 }
 @book{baumgartner2013applications,
   title={Applications of the Monte Carlo method in statistical physics},
@@ -218,6 +223,8 @@ month = "Oct",
 booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific Insight}},
 }
 @misc{memc,
-  key={Monte-Carlo for elastic shells},
+  author={Vipin Agrawal, Vikash Pandey and Dhrubaditya Mitra},
+  title={MeMC: A package for monte-carlo simulations of spherical shells},
+  year={2022},
   howpublished="\url{https://github.com/vipinagrawal25/MeMC}"
 }
