### 2022-05-30 #4305 @victorapm
Hi @vipinagrawal25 and @thecalculon, congratulations for your nice work! I have finished my review, the most important points I wanted to highlight are code organization and performance. Please find all my comments below, and let me know if you have questions. Thank you!

- [ ] Consider fixing the compilation warnings listed below:
```
src/init.cpp: In function ‘void init_read_parameters(MBRANE_para*, AFM_para*, MCpara*, char*)’:
src/init.cpp:102:37: warning: format ‘%d’ expects a matching ‘int*’ argument [-Wformat=]
  102 |         sscanf(buff,"%lf %lf %d %d %d", &td1, &td2,  &t_n2, &t_n3);
      |                                    ~^
      |                                     |
      |                                     int*
src/init.cpp:80:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   80 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:81:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   81 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:82:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   82 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:90:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   90 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:91:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   91 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:99:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   99 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:100:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
  100 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:101:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
  101 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:108:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
  108 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:109:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
  109 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/init.cpp:110:14: warning: ignoring return value of ‘char* fgets(char*, int, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
  110 |         fgets(buff,255,(FILE*)f2);
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~
src/forces_surf.cpp: In function ‘Vec3d determine_xyz_parabola(Vec3d, AFM_para)’:
src/forces_surf.cpp:45:20: warning: suggest parentheses around comparison in operand of ‘&’ [-Wparentheses]
   45 |         if(fabs(y0)< 1e-15 & fabs(z0)< 1e-15 ){
      |            ~~~~~~~~^~~~~~~
src/forces_surf.cpp:28:9: warning: variable ‘nroot’ set but not used [-Wunused-but-set-variable]
   28 |     int nroot;
      |         ^~~~~
src/hdf5_io.cpp: In function ‘void io_read_config(double*, int, char*)’:
src/hdf5_io.cpp:86:10: warning: ignoring return value of ‘size_t fread(void*, size_t, size_t, FILE*)’ declared with attribute ‘warn_unused_result’ [-Wunused-result]
   86 |     fread(Pos, N*sizeof(double), 1, fid);
      |     ~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```
- [ ] Is there a reason to use 60000 MC steps by default? Should this parameter be
      accessible to the user in the `exe_start` driver?
- [ ] Code organization:
1. Consider generating a library so that external application codes can link to it.
1. Consider moving the bulk of the work performed in the drivers `start.cpp` and
   `memc.cpp` to separate functions located at the `src/` folder. This way, you can create
   a library that distributes those functionalities to external users.
1. I could not find API documentation anywhere. For example, what are the meanings of the
   input/output parameters used in functions such as `monte_carlo_surf2d`,
   `pairlj_total_energy`, and `monte_carlo_3d`? Consider using a documentation generator
   tool such as Doxygen. Note that functions should be annotated to be able to
   automatically generate the documentation.
1. Add a `make clean` target (could be the same as `make distclean`).
- [ ] Code performance:
1. Have you considered multithreading, e.g., via OpenMP, to speedup functions that take
   most of the execution time? For example, Gprof indicates that `cal_length` takes up to
   70% of the execution time of `exe_start` and `bending_energy_ipart` takes up to 90% of
   the execution time of `exe_memc`. Would it be possible to reduce the execution time of
   these particular functions?
- [ ] Minor changes in the text:
1. Lines 60-61: did you mean: **there are two elastic constants to be determined**? Can you
   elaborate on this statement?
1. Line 64: should the definition of **R** be moved closer to its first appearance, i.e.,
   closer to Equation 3?
1. What is the meaning of K_B in Equation 6?
1. What is the meaning of H in Equation 9?
1. Lines 110-113, 119, 127: which subsubsection are you talking about? I cannot see the
   numbers.
1. In Equation 17, how `\Delta V` is computed? Is it simply `V - V_0`?
1. Line 121-122: remove the text **::: subequations :::**
- [ ] Add community guidelines which should contain information about how to:
1. Contribute to the software.
1. Report issues or problems with the software.
1. Seek support.
- [ ] Add white background to the figure: https://github.com/vipinagrawal25/MeMC/blob/main/paper/fig/describe_theta.png
- [ ] Update links in https://github.com/vipinagrawal25/MeMC/blob/main/README.md#example
1. gnuplot: http://www.gnuplot.info/
1. GNU GSL: https://www.gnu.org/software/gsl/
