### 2022-05-30 566a6a3
small changes in paper.md
