+++ paper/paper.bib
@@ -223,7 +223,7 @@ month = "Oct",
 booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific Insight}},
 }
 @misc{memc,
-  author={Vipin Agrawal, Vikash Pandey and Dhrubaditya Mitra},
+  author={Vipin, Agrawal and Vikash, Pandey and Dhrubaditya, Mitra},
   title={MeMC: A package for monte-carlo simulations of spherical shells},
   year={2022},
   howpublished="\url{https://github.com/vipinagrawal25/MeMC}"
