+++ src/Metropolis.cpp
@@ -8,10 +8,18 @@
 std::mt19937 rng;
 
 void init_rng(uint32_t seed_val){
+
+	 ///  @brief Generates random number 
     rng.seed(seed_val);
 }
 
 bool Metropolis(double DE, double kbt ){
+    /// @brief Metropolis algorithm
+    /// @param DE change in energy
+    /// @param kbt boltzmann constant times temperature
+    /// @return True if DE< 0 or the random number generated is less than
+    /// exp(-DE/kbt)
+    /// @details see https://en.wikipedia.org/wiki/Metropolis%E2%80%93Hastings_algorithm 
     bool yes;
     double rand;
     std::uniform_real_distribution<> rand_real(0, 1);
@@ -26,6 +34,8 @@ bool Metropolis(double DE, double kbt ){
 
 double rand_inc_theta(double th0, 
         double dfac){
+    /// @brief increment the polar angle randomly  
+    /
     double dth;
     double tmp_th0;
     std::uniform_real_distribution<> rand_real(-1, 1);
@@ -40,9 +50,23 @@ double rand_inc_theta(double th0,
 }
 
 double energy_mc_3d(Vec3d *pos, MESH mesh, 
-        double *lij_t0, bool *is_attractive, int idx, bool *is_be_pos,
+        double *lij_t0, bool *is_attractive, int idx,
         MBRANE_para mbrane, 
         MCpara mcpara, AFM_para afm){
+
+    /// @brief Estimate the contribution from all the energies when a particle is moved randomly 
+    ///  @param Pos array containing co-ordinates of all the particles
+    ///  @param mesh mesh related parameters -- connections and neighbours
+    ///information
+    /// @param lij_t0 initial distance between points of membrane
+    /// @param is_attractive true if the zz sees the  bottom wall 
+    ///  @param idx index of ith particle which is moved;
+    ///  @param mbrane  Membrane related parameters;
+    /// @param mcpara Monte-Carlo related parameters
+    /// @param AFM afm related parameter 
+    /// @return Change in Energy when idx particle is moved
+
+
     double E_b, E_s;
     double E_stick;
     double  E_afm;
@@ -67,7 +91,6 @@ double energy_mc_3d(Vec3d *pos, MESH mesh,
             mbrane.pos_bot_wall, mbrane.epsilon, mbrane.sigma);
 
     E_afm = lj_afm(pos[idx], afm);
-    *is_be_pos = E_b > 0;
 
     return E_b + E_s + E_stick + E_afm;
 }
@@ -76,6 +99,18 @@ int monte_carlo_3d(Vec3d *pos, MESH mesh,
                 double *lij_t0, bool *is_attractive, 
                 MBRANE_para mbrane, 
                 MCpara mcpara, AFM_para afm){
+
+    /// @brief Monte-Carlo routine for the membrane 
+    ///  @param Pos array containing co-ordinates of all the particles
+    ///  @param mesh mesh related parameters -- connections and neighbours
+    ///information
+    /// @param lij_t0 initial distance between points of membrane
+    /// @param is_attractive true if the zz sees the  bottom wall 
+    ///  @param mbrane  Membrane related parameters;
+    /// @param mcpara Monte-Carlo related parameters
+    /// @param AFM afm related parameter 
+    /// @return number of accepted moves 
+
     int i, move;
     int num_nbr, cm_idx;
     double x_o, y_o, z_o, x_n, y_n, z_n;
@@ -84,7 +119,6 @@ int monte_carlo_3d(Vec3d *pos, MESH mesh,
     double vol_i, vol_f;
[…]
