+++ src/forces_surf.cpp
@@ -4,6 +4,17 @@
 #define sign(x) ((x > 0) ? 1 : ((x < 0) ? -1 : 0))
 
 double cotangent(Vec3d si, Vec3d sk, Vec3d sj){
+
+  	 ///
+	 ///  @param si  coordinate of ith point
+	 ///  @param sk  coordinate of kth point
+	 ///  @param sj  coordinate of jth point 
+     
+	 ///  @return   ({si-sk}.{sj-sk})/sqrt(({si-sk}x{sj-sk})^2)
+     /// angle between vector si and sj
+	 ///
+
+
     Vec3d drik, drjk, cross;
     double cot_theta;  
     double inner_prod;
@@ -18,6 +29,10 @@ double cotangent(Vec3d si, Vec3d sk, Vec3d sj){
 }
 //
 double cotangent(double a, double b, double c){
+
+  	 /// @brief  a, b, c are the length of the sides of a triangle 
+     
+	 ///  @return   0.25*(a*a+b*b-c*c)/area; where area is the area of triangle
     double s = 0.5*(a+b+c);
     double area = sqrt(s*(s-a)*(s-b)*(s-c));
     double cot_theta=0.25*(a*a+b*b-c*c)/area;
@@ -25,6 +40,15 @@ double cotangent(double a, double b, double c){
 }
 
 Vec3d determine_xyz_parabola(Vec3d pos, AFM_para afm) {
+    ///
+	 ///  @param pos  position of a point in membrane
+	 ///  @param afm  parameters of AFM tip
+     //
+     /// 
+	 ///  @return  position of point on afm tip nearest to pos 
+	 ///
+
+
     int nroot;
     double a, b, c, d;
     double a0, c0;
@@ -75,6 +99,14 @@ Vec3d determine_xyz_parabola(Vec3d pos, AFM_para afm) {
 double volume_ipart(Vec3d *pos, 
         int *node_nbr, 
         int num_nbr, int idx, MBRANE_para para){
+     /// @brief Estimate the volume substended by voronoi area of the ith particle
+	 ///  @param Pos array containing co-ordinates of all the particles
+	 ///  @param idx index of ith particle;
+	 ///  @param node_nbr nearest neigbours of idx; 
+	 ///  @param num_nbr number of nearest neigbours of idx; 
+	 ///  @param para  Membrane related parameters;
+     /// @return Volume substended by ith particle.
+
 
     int i, j, k;
     double volume1;
@@ -105,6 +137,17 @@ double volume_ipart(Vec3d *pos,
 double stretch_energy_ipart(Vec3d *pos, 
         int *node_nbr, double *lij_t0,
         int num_nbr, int idx, MBRANE_para para){
+
+    /// @brief Estimate the Stretching energy contribution when ith particle position changes
+    ///  @param Pos array containing co-ordinates of all the particles
+    ///  @param idx index of ith particle;
+    ///  @param node_nbr nearest neigbours of idx; 
+    /// @param lij_t0 initial distance between points of membrane
+    ///  @param num_nbr number of nearest neigbours of idx; 
+    ///  @param para  Membrane related parameters;
+    /// @return Stretching Energy contribution when ith particle is displaced 
+
+
     //
     double HH;
     double idx_ener;
@@ -127,7 +170,19 @@ double stretch_energy_ipart(Vec3d *pos,
 //
 //Q. Given two cotangent angles, it returns either the area due to perpendicular bisector,
 // or the barycenter.
-double voronoi_area(double cotJ, double cotK, double jsq, double ksq, double area){
+double voronoi_area(double cotJ, double cotK, 
+        double jsq, double ksq, double area){
[…]

[…]
