+++ docs/latex/structAFM__para.tex
@@ -1,0 +1,62 @@
+\hypertarget{structAFM__para}{}\doxysection{AFM\+\_\+para Struct Reference}
+\label{structAFM__para}\index{AFM\_para@{AFM\_para}}
+
+
+{\ttfamily \#include $<$global.\+h$>$}
+
+\doxysubsection*{Public Attributes}
+\begin{DoxyCompactItemize}
+\item 
+double \mbox{\hyperlink{structAFM__para_ad7f167d6a1019712aeabd9997c9dcb49}{tip\+\_\+pos\+\_\+z}}
+\begin{DoxyCompactList}\small\item\em Afm structure. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{structAFM__para_a60fb8fe4984a3b201c6daeeae9394746}{tip\+\_\+rad}}
+\item 
+double \mbox{\hyperlink{structAFM__para_aee2a2fc662def1a1563e8b45218ec351}{sigma}}
+\item 
+double \mbox{\hyperlink{structAFM__para_aee9ead7ceada6db7344f0f914806518c}{epsilon}}
+\end{DoxyCompactItemize}
+
+
+\doxysubsection{Member Data Documentation}
+\mbox{\Hypertarget{structAFM__para_aee9ead7ceada6db7344f0f914806518c}\label{structAFM__para_aee9ead7ceada6db7344f0f914806518c}} 
+\index{AFM\_para@{AFM\_para}!epsilon@{epsilon}}
+\index{epsilon@{epsilon}!AFM\_para@{AFM\_para}}
+\doxysubsubsection{\texorpdfstring{epsilon}{epsilon}}
+{\footnotesize\ttfamily double AFM\+\_\+para\+::epsilon}
+
+\mbox{\Hypertarget{structAFM__para_aee2a2fc662def1a1563e8b45218ec351}\label{structAFM__para_aee2a2fc662def1a1563e8b45218ec351}} 
+\index{AFM\_para@{AFM\_para}!sigma@{sigma}}
+\index{sigma@{sigma}!AFM\_para@{AFM\_para}}
+\doxysubsubsection{\texorpdfstring{sigma}{sigma}}
+{\footnotesize\ttfamily double AFM\+\_\+para\+::sigma}
+
+\mbox{\Hypertarget{structAFM__para_ad7f167d6a1019712aeabd9997c9dcb49}\label{structAFM__para_ad7f167d6a1019712aeabd9997c9dcb49}} 
+\index{AFM\_para@{AFM\_para}!tip\_pos\_z@{tip\_pos\_z}}
+\index{tip\_pos\_z@{tip\_pos\_z}!AFM\_para@{AFM\_para}}
+\doxysubsubsection{\texorpdfstring{tip\_pos\_z}{tip\_pos\_z}}
+{\footnotesize\ttfamily double AFM\+\_\+para\+::tip\+\_\+pos\+\_\+z}
+
+
+
+Afm structure. 
+
+
+\begin{DoxyParams}{Parameters}
+{\em tip\+\_\+pos\+\_\+z;} & position of tip in z \\
+\hline
+{\em tip\+\_\+rad;} & radius of the tip \\
+\hline
+{\em sigma,epsilon;} & sigma epsilon of the LJ potential used to model AFM \\
+\hline
+\end{DoxyParams}
+\mbox{\Hypertarget{structAFM__para_a60fb8fe4984a3b201c6daeeae9394746}\label{structAFM__para_a60fb8fe4984a3b201c6daeeae9394746}} 
+\index{AFM\_para@{AFM\_para}!tip\_rad@{tip\_rad}}
+\index{tip\_rad@{tip\_rad}!AFM\_para@{AFM\_para}}
[…]
