+++ docs/latex/structMBRANE__para.tex
@@ -1,0 +1,170 @@
+\hypertarget{structMBRANE__para}{}\doxysection{MBRANE\+\_\+para Struct Reference}
+\label{structMBRANE__para}\index{MBRANE\_para@{MBRANE\_para}}
+
+
+{\ttfamily \#include $<$global.\+h$>$}
+
+\doxysubsection*{Public Attributes}
+\begin{DoxyCompactItemize}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_a3b3208469c3975907ee52245aafbb1af}{coef\+\_\+bend}}
+\begin{DoxyCompactList}\small\item\em Membrane Structure. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{structMBRANE__para_affc9e74002863428d7db15a1595ebef0}{coef\+\_\+str}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_aff393fb70e183bb07499ec828a0f57d8}{coef\+\_\+vol\+\_\+expansion}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_a64535a6612254b623c8f6a867bc3c197}{radius}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_adf0a7897cc9709acdbefed21aac7f275}{sigma}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_aa09bfd0d69a21eaafd518480d83766c2}{epsilon}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_a4cfca88519586156c3e8da03e304a722}{theta}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_a208dcc5adc94587c7467b7450089c668}{sp\+\_\+curv}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_a7f2cff2530979a3be0b51e78d84504b5}{pos\+\_\+bot\+\_\+wall}}
+\item 
+double \mbox{\hyperlink{structMBRANE__para_a8bedadd7cc3c6ee53f3c01c01c3d92ad}{av\+\_\+bond\+\_\+len}}
+\item 
+double $\ast$ \mbox{\hyperlink{structMBRANE__para_abeab3136ffef36abdaa27fdbd3bb3f4b}{tot\+\_\+energy}}
+\item 
+double $\ast$ \mbox{\hyperlink{structMBRANE__para_aae53f901fd0b69ae49e5c4ecbf5b0285}{volume}}
+\item 
+int \mbox{\hyperlink{structMBRANE__para_a49aaab499b5e5f74564cd71a3552d50e}{N}}
+\item 
+int \mbox{\hyperlink{structMBRANE__para_afda8ed2029922bbffddbc4f9682e1974}{num\+\_\+triangles}}
+\item 
+int \mbox{\hyperlink{structMBRANE__para_a5e8651cd5c9f80c99f0694d990c79050}{num\+\_\+nbr}}
+\end{DoxyCompactItemize}
+
+
+\doxysubsection{Member Data Documentation}
+\mbox{\Hypertarget{structMBRANE__para_a8bedadd7cc3c6ee53f3c01c01c3d92ad}\label{structMBRANE__para_a8bedadd7cc3c6ee53f3c01c01c3d92ad}} 
+\index{MBRANE\_para@{MBRANE\_para}!av\_bond\_len@{av\_bond\_len}}
+\index{av\_bond\_len@{av\_bond\_len}!MBRANE\_para@{MBRANE\_para}}
+\doxysubsubsection{\texorpdfstring{av\_bond\_len}{av\_bond\_len}}
+{\footnotesize\ttfamily double MBRANE\+\_\+para\+::av\+\_\+bond\+\_\+len}
+
[…]
