+++ docs/latex/structVec2d.tex
@@ -1,0 +1,33 @@
+\hypertarget{structVec2d}{}\doxysection{Vec2d Struct Reference}
+\label{structVec2d}\index{Vec2d@{Vec2d}}
+
+
+{\ttfamily \#include $<$Vector.\+h$>$}
+
+\doxysubsection*{Public Attributes}
+\begin{DoxyCompactItemize}
+\item 
+double \mbox{\hyperlink{structVec2d_a2f2289c32bb0ae9c85facaf6b4643fa8}{x}}
+\item 
+double \mbox{\hyperlink{structVec2d_aa314dea89fb8f30bc110f184d0a2813c}{y}}
+\end{DoxyCompactItemize}
+
+
+\doxysubsection{Member Data Documentation}
+\mbox{\Hypertarget{structVec2d_a2f2289c32bb0ae9c85facaf6b4643fa8}\label{structVec2d_a2f2289c32bb0ae9c85facaf6b4643fa8}} 
+\index{Vec2d@{Vec2d}!x@{x}}
+\index{x@{x}!Vec2d@{Vec2d}}
+\doxysubsubsection{\texorpdfstring{x}{x}}
+{\footnotesize\ttfamily double Vec2d\+::x}
+
+\mbox{\Hypertarget{structVec2d_aa314dea89fb8f30bc110f184d0a2813c}\label{structVec2d_aa314dea89fb8f30bc110f184d0a2813c}} 
+\index{Vec2d@{Vec2d}!y@{y}}
+\index{y@{y}!Vec2d@{Vec2d}}
+\doxysubsubsection{\texorpdfstring{y}{y}}
+{\footnotesize\ttfamily double Vec2d\+::y}
+
+
+
+The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
+\item 
+includes/\mbox{\hyperlink{Vector_8h}{Vector.\+h}}\end{DoxyCompactItemize}
