+++ search/variables_3.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['epsilon_0',['epsilon',['../structMBRANE__para.html#aa09bfd0d69a21eaafd518480d83766c2',1,'MBRANE_para::epsilon()'],['../structAFM__para.html#aee9ead7ceada6db7344f0f914806518c',1,'AFM_para::epsilon()'],['../structLJpara.html#a0103369ba9ebdfbe949b9aa6a96fc702',1,'LJpara::epsilon()']]]
+];
