### 2022-06-13 8b991ce
output dir from docs to .
