+++ paper/prof_report.txt
@@ -1,12 +1,635 @@
-Original code -> 
-	total run time = 157 sec.
-	85.9%(53.60+81.22se) in bending_energy_ipart
-	mc_total_iters=1e4.
-	report in analysis0.txt
-commenting line 155-157:
-	total run time = 153.56 sec.
-	85.7% (50.38s+81.21s) in bending_energy_ipart
-	report in analysis1.txt
-Use overloaded parameters instead of inline functions:
-	83%(47.52+80.81s)
-89% of the time is spent in bending_energy_ipart function for mc_total_iters>1e4.
+## Membrane parameters
+N coef_bending  coef_stretching coef_vol_expansion sp_curve
+512  2.50          25.00            1000000.00        2.0
+radius  pos_bot_wall  sigma epsilon   theta_attractive
+1.00     -1.05          0.17     4.00     0.52
+## Montecarlo parameters
+Dfac  kBT mc_total_iters  mc_dump_iter
+4       1.00  1000            100
+## Afm Tip parameters
+tip_radius  tip_pos_z afm_sigma afm_epsilon
+0.20         1.05       0.17         4.00
+
+compiled with flag '-pg'.
+After the compilation, run the code with para_file (given above).
+This will generate gmon.out in the respected directory.
+To get an analysis report use : gprof /path/to/exe_memc gmon.out > analysis.txt
+
+after compilation run from example folder ../bin/exe_memc para_file.in out
+
+Flat profile:
+
+Each sample counts as 0.01 seconds.
+  %   cumulative   self              self     total           
+ time   seconds   seconds    calls  ms/call  ms/call  name    
+ 33.13      4.87     4.87 14288498     0.00     0.00  bending_energy_ipart(Vec3d*, int*, int, int, MBRANE_para)
+ 15.00      7.08     2.21 1268284713     0.00     0.00  Vec3d::Vec3d()
+ 11.63      8.79     1.71 530110262     0.00     0.00  Vec3d::operator-(Vec3d)
+  7.96      9.96     1.17 270153928     0.00     0.00  cross_product(Vec3d, Vec3d)
+  5.44     10.76     0.80 740825952     0.00     0.00  inner_product(Vec3d, Vec3d)
+  4.59     11.43     0.68 233156330     0.00     0.00  Vec3d_add(Vec3d, Vec3d, double)
+  4.15     12.04     0.61 171940588     0.00     0.00  voronoi_area(double, double, double, double, double)
+  3.74     12.59     0.55 272199380     0.00     0.00  norm(Vec3d)
+  1.97     12.88     0.29  2048512     0.00     0.00  volume_ipart(Vec3d*, int*, int, int, MBRANE_para)
+  1.90     13.16     0.28 14288498     0.00     0.00  Vec3d::operator*(double)
+  1.53     13.39     0.23        1   225.00   225.00  _GLOBAL__sub_I__Z22init_system_random_posP5Vec2ddiPc
+  1.22     13.57     0.18 12243046     0.00     0.00  std::abs(double)
+  1.02     13.72     0.15  2048512     0.00     0.00  stretch_energy_ipart(Vec3d*, int*, double*, int, int, MBRANE_para)
[…]
