+++ Examples/analysis.txt
@@ -1,616 +1,0 @@
-Flat profile:
-
-Each sample counts as 0.01 seconds.
-  %   cumulative   self              self     total           
- time   seconds   seconds    calls  ms/call  ms/call  name    
- 33.13      4.87     4.87 14288498     0.00     0.00  bending_energy_ipart(Vec3d*, int*, int, int, MBRANE_para)
- 15.00      7.08     2.21 1268284713     0.00     0.00  Vec3d::Vec3d()
- 11.63      8.79     1.71 530110262     0.00     0.00  Vec3d::operator-(Vec3d)
-  7.96      9.96     1.17 270153928     0.00     0.00  cross_product(Vec3d, Vec3d)
-  5.44     10.76     0.80 740825952     0.00     0.00  inner_product(Vec3d, Vec3d)
-  4.59     11.43     0.68 233156330     0.00     0.00  Vec3d_add(Vec3d, Vec3d, double)
-  4.15     12.04     0.61 171940588     0.00     0.00  voronoi_area(double, double, double, double, double)
-  3.74     12.59     0.55 272199380     0.00     0.00  norm(Vec3d)
-  1.97     12.88     0.29  2048512     0.00     0.00  volume_ipart(Vec3d*, int*, int, int, MBRANE_para)
-  1.90     13.16     0.28 14288498     0.00     0.00  Vec3d::operator*(double)
-  1.53     13.39     0.23        1   225.00   225.00  _GLOBAL__sub_I__Z22init_system_random_posP5Vec2ddiPc
-  1.22     13.57     0.18 12243046     0.00     0.00  std::abs(double)
-  1.02     13.72     0.15  2048512     0.00     0.00  stretch_energy_ipart(Vec3d*, int*, double*, int, int, MBRANE_para)
-  0.85     13.84     0.13  3904219     0.00     0.00  double std::generate_canonical<double, 53ul, std::mersenne_twister_engine<unsigned long, 32ul, 624ul, 397ul, 31ul, 2567483615ul, 11ul, 4294967295ul, 7ul, 2636928640ul, 15ul, 4022730752ul, 18ul, 1812433253ul> >(std::mersenne_twister_engine<unsigned long, 32ul, 624ul, 397ul, 31ul, 2567483615ul, 11ul, 4294967295ul, 7ul, 2636928640ul, 15ul, 4022730752ul, 18ul, 1812433253ul>&)
-  0.75     13.95     0.11 28576996     0.00     0.00  Vec3d::operator/(double)
-  0.75     14.06     0.11  7808438     0.00     0.00  std::log(long double)
-  0.68     14.16     0.10  2048000     0.00     0.01  energy_mc_3d(Vec3d*, MESH, double*, bool*, int, MBRANE_para, MCpara, AFM_para)
-  0.48     14.23     0.07  2048000     0.00     0.01  bending_energy_ipart_neighbour(Vec3d*, MESH, int, MBRANE_para)
-  0.44     14.30     0.07  8832438     0.00     0.00  std::mersenne_twister_engine<unsigned long, 32ul, 624ul, 397ul, 31ul, 2567483615ul, 11ul, 4294967295ul, 7ul, 2636928640ul, 15ul, 4022730752ul, 18ul, 1812433253ul>::operator()()
-  0.41     14.36     0.06    14155     0.00     0.00  std::mersenne_twister_engine<unsigned long, 32ul, 624ul, 397ul, 31ul, 2567483615ul, 11ul, 4294967295ul, 7ul, 2636928640ul, 15ul, 4022730752ul, 18ul, 1812433253ul>::_M_gen_rand()
[…]
