+++ paper/response.txt
@@ -1,3 +1,5 @@
+We thank the referees for the generous comments. We have taken care of most of them. The comments have significantly improved the quality of the code. Below, we summarize the changes that we have made in our package as well as in the paper.
+
 `Consider fixing the compilation warnings listed below:
 src/init.cpp: In function ‘void init_read_parameters(MBRANE_para*, AFM_para*, MCpara*, char*)’:
 src/init.cpp:102:37: warning: format ‘%d’ expects a matching ‘int*’ argument [-Wformat=] ...`
@@ -32,7 +34,7 @@ pairlj_total_energy, and monte_carlo_3d? Consider using a documentation generato
 tool such as Doxygen. Note that functions should be annotated to be able to
 automatically generate the documentation.`
 
-We have generated the documentation using Doxygen and is now hosted at vipinagrawal25.github.io/memc.
+We have generated the documentation using Doxygen and is now hosted at vipinagrawal25.github.io/MeMC. We have mentioned this in readme as well.
 
 `Add a make clean target (could be the same as make distclean).`
 
@@ -46,57 +48,38 @@ most of the execution time? For example, Gprof indicates that cal_length takes u
 the execution time of exe_memc. Would it be possible to reduce the execution time of
 these particular functions?`
 
-We thank the referee for the comment. The number of calls to `cal_length` in
-`exe_start` is considerably large. Therefore, we expect this function to be the
-bottleneck. The sole purpose of `exe_start` is to generate initial configuration to be
-fed into `exe_memc`. A single simulation can generate multiple such configurations
-and the user need not execute `exe_start` again. Thus, optimizing this driver is not
-beneficial. 
-
-The second comment regarding optimizing `exe_memc` was useful. We were able to optimize the
-`bending_energy_ipart` further by vectorizing some operations and reduce the compute time by 4%. 
-However, this routine is called the atleast 6 times more than any other functions. 
-and the number of floating point operations in the function is significantly large.
+We thank the referee for this suggestion. Regarding optimizing `exe_memc`: we are able to optimize the `bending_energy_ipart` further by vectorizing some operations and reducing the compute time by 4%. 
+However, this routine is called the atleast 6 times more than any other functions, 
+and the number of floating point operations in the function is large.
 It is natural to expect this function to be a major bottleneck in the code.
-Although, we have included the current profile report in the repository (paper/prof_report.txt).
+The current profile report is included in the repository (paper/prof_report.txt).
 
-We have not considered multi-threading as of now. In the Monte-Carlo simulations, 
-the parallelization is not straightforward as only one particle can be moved
-randomly at a given instance. 
-During parallelization it may happen that multiple moves are made simultaneously
-which would
-lead to errors. Although, as suggested, we agree that certain routines can be made parallel 
-using OpenMP. 
-The current problems that we are trying to solve using the code requires a lot of ensemble
-averaging. Therefore, parallelization using OpenMP Pragma is not beneficial for us
-at this moment. We have added the suggestion in the todo list and will test them in future.
+The number of calls to `cal_length` in `exe_start` is considerably large. Therefore, we expect this function to be the bottleneck too. However, the sole purpose of `exe_start` is to generate initial configuration to be fed into `exe_memc`. A single simulation can generate multiple such configurations and the user need not execute `exe_start` again. Thus, optimizing this driver is not so urgent.
+
+We agree with the referee that openMP parallelization can be helpful. However, in the Monte-Carlo simulations, the parallelization is not immediately obvious as only one particle can be moved
[…]
