+++ paper/paper.bib
@@ -39,17 +39,15 @@ title={Competition between bending and internal pressure governs the mechanics o
   publisher={APS},
   doi={10.1103/physrevx.7.011002}
 }
-
 @article{hw2002stomatocyte,
   title={Stomatocyte--discocyte--echinocyte sequence of the human red blood cell: Evidence for the bilayer--couple hypothesis from membrane mechanics},
-  author={HW, Gerald Lim and Wortis, Michael and Mukhopadhyay, Ranjan},
+  author={Lim HW, Gerald and Wortis, Michael and Mukhopadhyay, Ranjan},
   journal={Proceedings of the National Academy of Sciences},
   volume={99},
   number={26},
   pages={16766--16769},
   year={2002},
-  publisher={National Acad Sciences},
-  doi={10.1073/pnas.202617299}
+  publisher={National Acad Sciences}
 }
 @misc{comsol,
   author={COMSOL Inc.},
