+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "MeMC: A package for monte-carlo simulations of spherical shells"
+title: "MeMC: A package for Monte Carlo simulations of spherical shells"
 tags:
 authors:
   - name: Vipin Agrawal
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 # Summary
 
-  The MeMC is an open-source software package for monte-carlo simulation
+  The MeMC is an open-source software package for Monte Carlo simulation
   of elastic shells. It is designed as a tool to interpret the
   force-distance data generated by indentation of biological
   nano-vesicles by atomic force microscopes. The code is written in c++
@@ -75,7 +75,7 @@ elastic shell.
 There are commercial packages, e.g., COMSOL [@comsol], to calculate the
 force-distance curve of solid bodies and closed membranes with fluids
 inside under the action of external forces but to the best of our
-knowledge there is no package that includes thermal effects. Monte-carlo
+knowledge there is no package that includes thermal effects. Monte Carlo
 simulations of elastic membranes, that does include thermal
 fluctuations, have been done for more than three
 decades [@goetz1999mobility; @bowick2001universality; @auth2005fluctuation; @paulose2012fluctuating],
@@ -157,9 +157,9 @@ $\mbox{ET} \approx 10^3$.
 
 Following Ref. [@gompper2004triangulated], we use a triangulated-network
 grid in the following manner. We start with $N$ randomly chosen points
-on a sphere . Then, we run a Monte-Carlo simulation, with a Lennard-Jones
+on a sphere . Then, we run a Monte Carlo simulation, with a Lennard-Jones
 (LJ) repelling potential, of these points moving on the surface. Once
-the surface Monte-Carlo (SMC) has reached an equilibrium, we use the
+the surface Monte Carlo (SMC) has reached an equilibrium, we use the
 algorithm in Ref. [@caroli2009robust] to construct the Delaunay
 triangulation of these points. The connection between the points thus
 formed is kept unchanged. In the rest of this paper, we call this the
@@ -177,7 +177,7 @@ the initial configuration shown in (A). (C) An example of regular grid }
 
 ## Energy
 
-The basic algorithm of Monte-Carlo simulations is straightforward and
+The basic algorithm of Monte Carlo simulations is straightforward and
 well-known [see, e.g., @baumgartner2013applications]. We randomly choose
 a point on the grid and move it by a random amount. We calculate the
 change in energy due to this movement. We accept or reject the move by
@@ -415,7 +415,7 @@ In the github repository, we also provide a subdirectory called
 the user can run the code (without the AFM tip and the bottom plate). It
 takes almost $30$ minutes on Intel(R) Core(TM) i5-8265U CPU. The run
 produces a probability distribution function (PDF) of the total energy
-after $50,000$ number of monte-carlo steps. By running
+after $50,000$ number of monte carlo steps. By running
 `gnuplot plot.gnu` (this requires the software gnuplot) the user can
 compare the PDF obtained by their run with a PDF that we provide.
 
@@ -435,4 +435,5 @@ resources provided by the Swedish National Infrastructure for Computing
 The code is licensed under GPL-3.0 and hosted at [here](https://github.com/vipinagrawal25/MeMC).
 For any contribution, the developer can send a [pull request](https://github.com/vipinagrawal25/MeMC/pulls).
 For any queries, the user should open an [issue](https://github.com/vipinagrawal25/MeMC/issues) on github.
+
 # References
