### 2022-06-17 #4305 @fboehm
@vipinagrawal25 - can you examine the above output from the check references command? Is the suggested "missing doi" accurate? If so, would you please update your manuscript to include it?
