+++ paper/paper.bib
@@ -47,7 +47,8 @@ title={Competition between bending and internal pressure governs the mechanics o
   number={26},
   pages={16766--16769},
   year={2002},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.202617299}
 }
 @misc{comsol,
   author={COMSOL Inc.},
