### 2022-06-17 #4305 @thecalculon
> @vipinagrawal25 - can you examine the above output from the check references command? Is the suggested "missing doi" accurate? If so, would you please update your manuscript to include it?

@fboehm the doi suggeted by the 'good bot' was correct. We have updated it in the paper.
