### 2022-06-17 #4305 @fboehm
the doi for Gompper & Kroll resolves to a reference other than that listed. Can you fix this?
