+++ paper/paper.md
@@ -20,7 +20,7 @@ authors:
 affiliations:
   - name: Nordita, KTH Royal Institute of Technology and Stockholm University, Roslagstullsbacken 23, 10691 Stockholm, Sweden.
     index: 1
-  - name: Department of Physics, Stockholm university, Stockholm,Sweden.
+  - name: Department of Physics, Stockholm University, Stockholm,Sweden.
     index: 2
   - name: KTH Royal Institute of Technology, Sweden.
     index: 3
@@ -64,7 +64,7 @@ two important ways
 1.  The nano-vesicles are much smaller hence thermal fluctuations may
     effectively renormalize the elastic coefficients [@paulose2012fluctuating; @kovsmrlj2017statistical ].
 
-2.  Cell membranes are strongly coupled to an underlying cytoskelton.
+2.  Cell membranes are strongly coupled to an underlying cytoskeleton.
     Hence they may be modeled by a solid body [@hw2002stomatocyte] but
     nano-vesicles must be modeled as liquid filled elastic membranes.
 
@@ -76,11 +76,11 @@ There are commercial packages, e.g., COMSOL [@comsol], to calculate the
 force-distance curve of solid bodies and closed membranes with fluids
 inside under the action of external forces but to the best of our
 knowledge there is no package that includes thermal effects. Monte Carlo
-simulations of elastic membranes, that does include thermal
+simulations of elastic membranes, that includes thermal
 fluctuations, have been done for more than three
 decades [@goetz1999mobility; @bowick2001universality; @auth2005fluctuation; @paulose2012fluctuating],
-see also [@gompper2004triangulated for a review ]. But to the best of
-our knowledge there are no open-source code available. The goal of this
+see also [@gompper2004triangulated] for a review. But to the best of
+our knowledge there is no open-source code available. The goal of this
 package is to fill this gap in open-source software. 
 Our software is targeted towards physicists and biologists working in soft matter.
 
@@ -90,7 +90,7 @@ Our software is targeted towards physicists and biologists working in soft matte
 Our model of nano-vesicles is an amorphous membrane enclosing an
 incompressible fluid[@vorselen2017competition]. Unlike a solid ball, the
 force-distance relationship for such a model is linear for small
-deformation [@vorselen2017competition; @paulose2012fluctuating] if we
+deformation [@vorselen2017competition; @paulose2012fluctuating], if we
 ignore thermal fluctuations. Ref. [@vorselen2017competition] uses a
 similar model, ignoring thermal fluctuations, to interpret AFM
 measurement of nano-vesicles.
@@ -116,7 +116,7 @@ consider a fluid enclosed in a solid membrane, as done in
   u_{ij} = \frac{1}{2}\left( \partial_i u_j + \partial_j u_i + \partial_i w\partial_j w\right)
             -\delta_{ij}\frac{w}{R},
 \end{equation}
-where $R$ is the radius of the spherical cell, w$ is the out-of-plane deformation, and $\mathbf{u}$ is the
+where $R$ is the radius of the spherical cell, $w$ is the out-of-plane deformation, and $\mathbf{u}$ is the
 in-plane deformation, $p$ is the pressure, $\lambda$ and $\mu$ are the
 two in-plane Láme coefficients and $B$ is the bending modulus. The Láme
 coefficients are related to other elastic constant as follows [@LLelast]
@@ -167,12 +167,12 @@ formed is kept unchanged. In the rest of this paper, we call this the
 
 An alternative is to use a regular
 grid [@vliegenthart2006mechanical; @buenemann2008elastic]. This is
-achieved by approximating the sphere with geodesic polyhedron (\autoref{fig:grid}(C)).
[…]
