+++ paper/paper.bib
@@ -227,3 +227,14 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
   year={2022},
   howpublished="\url{https://github.com/vipinagrawal25/MeMC}"
 }
+@article{boek2005mechanical,
+  title={Mechanical properties of surfactant bilayer membranes from atomistic and coarse-grained molecular dynamics simulations},
+  author={Boek, ES and Padding, JT and den Otter, Wouter K and Briels, Willem J},
+  journal={The Journal of Physical Chemistry B},
+  volume={109},
+  number={42},
+  pages={19851--19858},
+  year={2005},
+  publisher={ACS Publications}
+  doi={10.1021/jp054372b}
+}
