+++ paper/paper.bib
@@ -235,6 +235,6 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
   number={42},
   pages={19851--19858},
   year={2005},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
   doi={10.1021/jp054372b}
 }
