+++ paper/paper.md
@@ -20,7 +20,7 @@ authors:
 affiliations:
   - name: Nordita, KTH Royal Institute of Technology and Stockholm University, Roslagstullsbacken 23, 10691 Stockholm, Sweden.
     index: 1
-  - name: Department of Physics, Stockholm University, Stockholm,Sweden.
+  - name: Department of Physics, Stockholm University, Stockholm, Sweden.
     index: 2
   - name: KTH Royal Institute of Technology, Sweden.
     index: 3
@@ -172,7 +172,7 @@ They are available from the *Meshzoo* library [@meshzoo]. In this paper
 we use $N=5120$ for the random grid and $N=5292$ for the regular grid.
 
 ![\textbf{Grid points on a sphere} (A)Triangulated random points on a sphere. (B) Triangulated points on sphere after $60000$ SMC iteration of
-the initial configuration shown in (A). (C) An example of regular grid
+the initial configuration shown in (A). (C) An example of regular grid.
 \label{fig:grid}](fig/spheres.pdf){width=100%}
 
 ## Energy
