+++ paper/paper.bib
@@ -91,13 +91,13 @@ title={Competition between bending and internal pressure governs the mechanics o
 }
 @inproceedings{gompper2004triangulated,
   title={Triangulated-surface models of fluctuating membranes},
-  author={Gompper, G and Kroll, DM},
+  author={Gompper, G and Kroll, D. M},
   booktitle={Statistical mechanics of membranes and surfaces},
   editor={Nelson, David and Piran, Tsvi and Weinberg, Steven},
   pages={359--426},
   year={2004},
   publisher={World Scientific},
-  doi={10.1103/physrevx.7.011002}
+  doi={10.1142/9789812565518_0012}
 }
 @book{LLelast,
   title={Theory of Elasticity},
