### 2022-06-20 #4305 @fboehm
The Gomper & Kroll ref's doi now resolves to the intended target.
