### 2022-06-25 #4305 @Kevin-Mattheus-Moerman
@vipinagrawal25 congratulations on getting this paper published in JOSS!

@fboehm thanks for editing this one!

And special thanks to @victorapm and @mgiardino for your review work!! :tada:
